requirements.txt
asgiref==3.4.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==2.1.0
Django==3.2.8
django-cors-headers==3.10.0
django-currentuser==0.5.3
django-filter==21.1
django-phonenumber-field==5.2.0
django-sequences==2.6
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.20.2
filelock==3.3.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jsonschema==4.1.2
MarkupSafe==2.0.1
numpy==1.21.4
oauthlib==3.1.1
phonenumbers==8.12.36
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.1
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
pyrsistent==0.18.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
typing_extensions==4.9.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.9.0
virtualenvwrapper-win==1.2.6
whitenoise==5.3.0
